hmlet/photos/flesh/pyproject.toml
[tool.poetry]
name = "photos"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
django = "^3.0.5"
furl = "^2.1.0"
djangorestframework = "^3.11.0"
django-filter = "^2.2.0"
Pillow = "^7.1.1"
requests = "^2.23.0"
djangorestframework_simplejwt = "^4.4.0"
django-minio-storage = "^0.3.7"
django-versatileimagefield = "^2.0"
fire = "^0.3.1"
loguru = "^0.4.1"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-factoryboy = "^2.0.3"
pytest-django = "^3.9.0"
django-environ = "^0.4.5"
pytest-socket = "^0.3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
